Add option to verify the latest backup
When -v or --verify is passed to the program, vintage backup
will compare the backed up files with the original files in the
user's directory by directly comparing contents. If there is a
mismatch, the file's modification time will be changed so that
it will be fully copied instead of linked during the next backup.