Quit if user folder changed from previous backup
Changing the source of user data to backup will result in none
of the files in a previous backup matching, which means that
all files will have to be copied anew. Plus, there will be nothing
to distinguish the different backup sets when browsing the
dated folders. The user should choose another backup location
to backup a different user folder.