Return a non-existent directory if no prior backup
Simplify the code by setting the last_backup_path to a
deleted temporary directory. This avoids constant checks
if the last_backup_path is None.  Since the directory cannot
exist (without deliberate sabotage), the filecmp.cmpfiles() result
will always be a list of errors, causing all files to be copied.