Remove process time from Timer
The time taken by this program does not happen within
this python code, but in the filecmp functions. The function
time.process_time() is useless here.

Extra timers revealed that the filecmp.cmpfiles() call takes up
almost all of the run time. Work continues to figure out
why it sometimes takes hours and other times minutes for
the same set of files.