Remove Timers
The cause of the slow down is filecmp.cmpfiles().

When shallow=True, the file contents are skipped when the
metadata (size, type, and modification time) are identical.
Otherwise, the size and file contents are checked.

To get a quick check like rsync, write a custom file comparator
that only looks at metadata with the same interface as
filecmp.cmpfiles().