Better use of temporary directory
Previously, there was a chance that the deleted temporary
directory could be recreated by another process, leading
to possible problems with the backup.

Now, the temporary directory is explicitly created for the
duration of the program, and the name of the folder is
explicitly checked against other paths so that its contents
are never accessed.

The modifications to the if __name__ == "__main__" block
ensure that the temporary folder is deleted.