Replace string paths with Path objects
This change allows for much less work to be done when doing operations
on paths. Specifically, comparing paths with different letter casing on
Windows. This is all handled behind the scenes, especially in functions
like filter_excluded_paths() which are now much shorter and simpler.

The Path class replaces all operations that create, modify, or perform
calculations on path. The functions in os.path only operate on strings,
so there's a danger of the path names "foo" and "FOO" being treated
differently. The os functions that are left (link, samefile, walk, etc.)
only take paths to query the file system, which doesn't require any
special handling.

Closes #21