Search for files and folders to recover in a directory (#28)
When a user doesn't know the name of a file or folder that they want to retrieve from backup, they can search through all the backups of a given folder for all files and folder names that have been backed up. After selecting the file from a list, it will be recovered as if by the --recover argument.

* Create function to list all backups

Also checks that path names are correct for backup. Use this new
function for find_previous_backup().

Closes #19

* Function to get backup source from backup location

* Make sure backup exists before recovery

* Search backups for files and folders to recover

Useful for when the name of a file or folder is unknown, but the
folder where it used to be is.

Closes #22