Unified include-exclude files (#44)
* Unify include and exclude files

Working solution to #41. Will consider if it would be better to filter
files on the fly rather than scanning whole directory before starting.

* Use argument to exclude symlinks

* Update docstrings

* Create function to find non-symlink directories

* Remove redundant Path() call

* Apply backup filters on the fly

This avoids having to iterate through the entire user folder to build up
a list of everything that will be backed up.

* Fix symlink handling

* Correct path in example

* Fix directory in help text

* Correctly handle glob patterns

There's no facility for testing recursively matching glob patterns of
directories without symlinks being followed. So, for each line of an
alter file, generate all paths that match and modify the list of all
files to be backed up as needed.

* Split long line

* Apply change set after adding everything

* Allow recursive (**) globs

* Document that order matters in the alter file

* More accurate argument names

* Fix tests for unified exclude-include file

* Fix docstrings

* Fix tests for alter files

* Fix tests for alter file version

* Put type definition near usage