Delete backups with readonly files
If a user backups up data that they have permission to read but not
tomodify (including delete) then they won't be able to delete old
backups. This is true even for files and folders in a Windows user's
home directory (some AppData subfolders, Contacts, Desktop, etc.).

With this change, when deleting a readonly file results in a
PermissionError, write permission is added to the file and the deletion
is retried. This should allow for clean deletions of old backups.

Finally, add a test that proves the problem and the solution.