Restore backup (#56)
This new function overwrites the files in the user's folder with the files in a chosen backup (--last-backup to use the most recent, --choose-backup to pick from a terminal menu). Unlike --recover, which copies individual files/folders and renames the recovered data to avoid clobbering user data, --restore overwrites all files with those in the backup. If there are files that are not present in the last backup, --keep-extra preserves them and --delete-extra deletes them. Note that --delete-extra not only deletes files that are newly created since the last backup, but also any files that were filtered out with --filter.

Closes #54.