New option: --hard-link-count
In order to limit the time a backup takes on Windows systems, randomly
copy files instead of hard-linking them when they haven't changed. The
probability of copying is specified by the new command line option
--hard-link-count. The files are chosen randomly for two reasons:

1) counting hard links when there are a lot of links takes a long time,
which is the problem we're trying to fix, and
2) It will limit the number of backups that need to be deleted for space
since each backup will have more copied data to delete.