New option: --hard-link-count (#68)
In order to limit the time a backup takes on Windows systems, randomly copy files instead of hard-linking them when they haven't changed. The probability of copying is specified by the new command line option --hard-link-count, which is interpreted as the average number of times a file is hard linked to previous backups before another copy is made. The files are chosen randomly because counting hard links when there are a lot of links takes a long time [1], which is the problem we're trying to fix.

Implements and closes #67

[1] See all the sources in issues #16 and #26.