Concurrency lock (#71)
Prevent multiple instances of Vintage Backup from operating on the same backup location. By default, the program ends immediately in case of conflict. The option --wait causes the program to wait for the other instance to finish.

Closes #65

* Lock to prevent simultaneous backup operations

* Delay locking as long as possible

Only user interaction, title printing, and the actual function should be
blocked by the lock file: user interaction so that the user doesn't do a
bunch of work only to be told to wait, title printing so the user
doesn't think something is happening when it's not, and the function to
prevent interference from other processes.