Make backup deletion tests more reliable
To avoid other programs writing to or deleting from the disk during a
test, only read the free space on the disk just before doing test backup
deletions. This reduces the amount of time during which writing to the
test hard drive can occur. This should reduce the likelihood of a false
failure.