Create class for on-the-fly filtering
Previously, the program would create a list of files for every line in
the filter file and add/remove files from a scan of all backup files
before starting the transfer of files. This could take several minutes.
Now, The backup can start immediately and every file is passed through
the filter file lines once to determine if it gets backed up.

The translation of filter file lines to glob patterns needs work. I
don't like the automatic appending of /** to any pattern that happens to
match a directory.