Document that filter lines only test files
Instead of trying to detect when a filter line refers to a directory,
document that only files are considered when filtering. The recursive
wildcard ** must be used to exclude directories by excluding all of
their contents. Update filter tests to reflect this.

Also, clean up filter tests to reduce redundancy and to make it easier
to add tests.