On-the-fly filtering (#76)
* Do not follow symlinks during backup deletion

* Create class for on-the-fly filtering

Previously, the program would create a list of files for every line in
the filter file and add/remove files from a scan of all backup files
before starting the transfer of files. This could take several minutes.
Now, The backup can start immediately and every file is passed through
the filter file lines once to determine if it gets backed up.

The translation of filter file lines to glob patterns needs work. I
don't like the automatic appending of /** to any pattern that happens to
match a directory.

* Simplify filter test

* Fix flake8 errors

* Remove unused variable

* Force filter lines for directories to end with **

* Update required Python version

* Fix filter documentation

* Change branch order to remove not in conditional

* Add tests for bad filter lines

* Document that filter lines only test files

Instead of trying to detect when a filter line refers to a directory,
document that only files are considered when filtering. The recursive
wildcard ** must be used to exclude directories by excluding all of
their contents. Update filter tests to reflect this.

Also, clean up filter tests to reduce redundancy and to make it easier
to add tests.

* Use str for default log argument

This is what setup_log_file() expects.

* Debugging log messages for filtering

* Exclude Windows junctions by default