Require typed confirmation to restore from backup
Restoring from backup is potentially the most destructive action this
program can take. That's because restoring from backup will overwrite
files and possibly delete other files. This is unlike recovering files
and folders, which do not overwrite existing files. So, in case there is
some mistake with the restoration command, show the user what action is
about to be taken and require the user to type "yes" before proceeding.