Handle stale locks
While a backup is running, the program will periodically write an
increasing number to the lock file. Any other Vintage Backup process can
check that the lock is still fresh by reading the file twice with a time
delay. If a process crashes without deleting the lock file (i.e., power
outage), then the next process can see that the lock is stale and delete
it.

Closes #83.