Delete unneeded directory existence check
This check was implemented when all paths to be backed up were found
before any backup started. There could be a long delay between finding a
directory and actually backing it up. Now that paths are backed up
immediately after being found and filtered, the delay is tiny. At worst,
some warnings will be logged regarding being unable to back up certain
files that were deleted in that fraction of a second delay.