Option to continue directory deletion after errors
Add an option--ignore_errors--to delete_directory_tree() that will skip
files and folders that cannot be deleted due to errors. This will allow
all other files and folders in that directory to be deleted and the
function to exit successfully. This is especially important when
deleting old backups so that the backup drive does not fill up due to
erroneous files preventing the deletion of old data.