Preview filter and shorter directory listing format (#154)
Two changes in this PR:

1. Add a `--preview-filter` function so users can see exactly which
files will make it through a filter.
2. Print the above file list in a shorter, easier to read format. Apply
this format to `--verify`, as well.

The short file list format is of the form:
```
/path/to/folder/
    file1.txt
    file2.txt
    ...
/path/to/folder/subfolder/
   file3.txt
   ...
```
and similarly for Windows paths. This format should be easier to read
since the file names are not way to the right--often beyond the edge of
the screen for deeply nested files.

Closes #149
Closes #152