Fix tests and fix log rotation
Fixed test by running a backup after the first log rotation. Previously,
the second backed up log file wouldn't be created because there was
nothing in the active log except "Start of new log file". So, there was
no logged backup to check.

Simplified log rotation by using the built-in RotatingFileHandler. There
is still a problem since the backupCount argument needs to be greater
than zero for rollover to work. Right now, this number is set to 1000 to
prevent backed up logs from being deleted. Some customization is
necessary to get rid of this arbitrary number.

Also, instead of using log timestamps, use backup datetimes to judge when to
rotate.