Manually close error log files on context exit
The assertLogs context restores the logs to the saved state they were in
prior to the context entrance by a simple copy of the logger list that
was saved. Since the error log is added after entry, this gets removed
without closing the file upon context exit. This new function closes all
file logs before exiting the assertLogs context.