Correct error in delete_too_frequent_backups()
The function parse_timespan_to_timepoint() returns a datetime in the
past. So, the proper comparison to determine if a backup should be
deleted is to compare the timestamp of the standard (the oldest backup)
with the past timestamp relative to the next backup.

This fix was found by the test that is also part of this commit.